Add DOI-based references to a bib file

Check if DOIs are present in a bib file passed as an argument, if one or more DOI are missing the associated references are imported and added to the bib file.

add_doi_reference(doi, bib = "./inSilecoRef.bib")

Arguments

doi

character vector of DOI to import.

bib

character path to master .bib file, default is inSilecoRef.bib, if the file does not exist it will be created.

Examples

# Example 1:
if (FALSE) {
add_doi_reference(doi = c(
  "10.1016/j.tree.2016.10.011", "10.1890/130230",
  "10.1515/9781400881376"
))
}